libalsa-intf: Modify aplay to consider file size in PCM header
The issue with aplay is that it was not reading the length in header.
when silence files are created using third party pcm generators, it
has some meta data at the end of file after all zeros. So aplay
will send that data also. This behavior is modified now to avoid
pops in the end of the file.

Change-Id: Ibd6d91503b95dc55e8cb119a23ccfcd27fb3fab4
CRs-Fixed: 324206